# openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/v1/Error.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "details": {
            "additionalProperties": {},
            "type": "object"
          },
          "status": {
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "type": "object"
      },
...
      "PageTask": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/v1/PageTask.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "next_cursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "Task": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/v1/Task.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "actor": {
            "enum": [
              "operator",
              "system"
            ],
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "environment_id": {
            "type": "string"
          },
          "finished_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "operation_id": {
            "type": "string"
          },
          "plan_hash": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "retry_of": {
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/TaskStep"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "target": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "deploy",
              "rollback",
              "backup",
              "backup_prune",
              "restore",
              "attach",
              "detach",
              "run",
              "script",
              "provision",
              "create",
              "update",
              "remove",
              "start",
              "stop",
              "destroy",
              "rotate"
            ],
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "workspace_type": {
            "enum": [
              "platform",
              "tenant"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "operation_id",
          "type",
          "target",
          "status",
          "workspace_type",
          "actor",
          "created_at",
          "updated_at",
          "started_at",
          "finished_at"
        ],
        "type": "object"
      },
      "TaskAccepted": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/v1/TaskAccepted.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "task_id": {
            "type": "string"
          }
        },
        "required": [
          "task_id"
        ],
        "type": "object"
      },
...
      "TaskStep": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Groundplane API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/activity": {
      "get": {
        "operationId": "activity.list",
        "parameters": [
          {
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "environment",
            "schema": {
              "pattern": "^env_[0-9A-HJKMNP-TV-Z]{26}$",
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "workspace",
            "schema": {
              "pattern": "^(platform|tnt_[0-9A-HJKMNP-TV-Z]{26})$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageTask"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List activity",
        "tags": [
          "Task"
        ]
      }
    },
...
    "/tasks": {
      "get": {
        "operationId": "task.list",
        "parameters": [
          {
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "environment",
            "schema": {
              "pattern": "^env_[0-9A-HJKMNP-TV-Z]{26}$",
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "workspace",
            "schema": {
              "pattern": "^(platform|tnt_[0-9A-HJKMNP-TV-Z]{26})$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageTask"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List tasks",
        "tags": [
          "Task"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "task.show",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^task_[0-9A-HJKMNP-TV-Z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Show a task",
        "tags": [
          "Task"
        ]
      }
    },
    "/tasks/{id}/abort": {
      "post": {
        "operationId": "task.abort",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^task_[0-9A-HJKMNP-TV-Z]{26}$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            },
            "description": "Accepted",
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Abort an in-flight task",
        "tags": [
          "Task"
        ]
      }
    },
...
    "/tasks/{id}/retry": {
      "post": {
        "operationId": "task.retry",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "pattern": "^task_[0-9A-HJKMNP-TV-Z]{26}$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAccepted"
                }
              }
            },
            "description": "Accepted",
            "headers": {
              "Content-Type": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Retry a failed task",
        "tags": [
          "Task"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}